requirements-minimal.txt
# FastAPI 및 서버
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0

# HTTP 클라이언트
httpx==0.26.0

# 최소 LangChain (Mock 모드용)
langchain-core==0.1.23
pydantic==2.5.3

# 테스트
pytest==7.4.4
pytest-asyncio==0.23.3
